requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.0.8
anilibria.py==0.3.3
async-timeout==4.0.3
attrs==23.2.0
backports.functools-lru-cache==2.0.0
beautifulsoup4==4.7.1
Brotli==1.0.9
certifi==2024.2.2
charset-normalizer==2.0.12
cheroot==6.5.4
cloudscraper==1.2.60
dateutils==0.6.6
deluge-client==1.7.0
enum34==1.1.6
falcon==1.4.1
feedparser==6.0.8
frozenlist==1.4.1
future==0.17.1
greenlet==2.0.1
html5lib==0.999999999
idna==3.7
itsdangerous==1.1.0
Mako==1.3.3
MarkupSafe==2.1.5
more-itertools==10.2.0
multidict==6.0.5
path==16.14.0
phpserialize==1.3
playwright==1.31.1
pyee==9.0.4
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-editor==1.0.4
python-mimeparse==1.6.0
pytz==2018.9
PyYAML==6.0.1
qbittorrent-api==2023.3.44
requests==2.27.1
requests-toolbelt==1.0.0
semver==2.8.1
sgmllib3k==1.0.0
six==1.12.0
soupsieve==2.5
SQLAlchemy==1.3.12
SQLAlchemy-Enum34==1.0.1
structlog==24.1.0
tinydb==4.8.0
transmission-rpc==7.0.5
transmissionrpc==0.11
typing_extensions==4.11.0
urllib3==1.26.18
webencodings==0.5.1
yarl==1.9.4
